# backend/cmd/api/docs/swagger.yaml
basePath: /api
definitions:
  internal_http_auth.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_http_auth.SignUpRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_http_category.CreateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  internal_http_category.RenameCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Stock management application API.
  title: Stockify API
  version: "1.0"
paths:
  /auth/me:
    get:
      description: Get authenticated user profile from JWT token
      produces:
      - application/json
      responses:
        "200":
          description: user data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/sign-in/email:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, returns JWT token via cookie
      parameters:
      - description: User credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_auth.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: invalid credentials
          schema:
            additionalProperties: true
            type: object
      summary: Sign in a user
      tags:
      - Auth
  /auth/sign-out:
    post:
      description: Clear JWT cookie to sign out
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Sign out
      tags:
      - Auth
  /auth/sign-up/email:
    post:
      consumes:
      - application/json
      description: Create a new user account with email, name, and password
      parameters:
      - description: User registration payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_auth.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: user_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties: true
            type: object
        "422":
          description: email already exists
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
  /category/:
    get:
      description: Retrieve all categories for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: list of categories
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get all categories
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: Create a new inventory category for the authenticated user
      parameters:
      - description: Category name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_category.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: category_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Create a new category
      tags:
      - Category
  /category/{id}:
    delete:
      description: Soft-delete a category and unassign its products
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: category_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Delete a category
      tags:
      - Category
    get:
      description: Retrieve a single category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: category data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid id
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Get a category by ID
      tags:
      - Category
    patch:
      consumes:
      - application/json
      description: Update the name of an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: New name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_http_category.RenameCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: category_id
          schema:
            additionalProperties: true
            type: object
        "400":
          description: validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - CookieAuth: []
      summary: Rename a category
      tags:
      - Category
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"